+++ cookbooks/mkdirs-nativex/attributes/default.rb
@@ -1,0 +1,8 @@
+# mkdirs-nativex path to directory being created
+default['mkdirs-nativex']['path'] = ["/automnt"]
+# user who owns the directory
+default['mkdirs-nativex']['owner'] = "root"
+# group that owns the directory
+default['mkdirs-nativex']['group'] = "root"
+# permissions mode
+default['mkdirs-nativex']['mode'] = "755"