+++ cookbooks/mkdirs-nativex/metadata.rb
@@ -1,0 +1,7 @@
+name             'mkdirs-nativex'
+maintainer       'NativeX'
+maintainer_email 'brett.stime@nativex.com'
+license          'All rights reserved'
+description      'Creates [empty] directories configured by a node attribute list'
+long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
+version          '0.1.0'